Create dedicated Attrable mixin for attr_ methods
