+++ Library/Homebrew/debrew.rb
@@ -1,6 +1,7 @@
 # typed: true
 # frozen_string_literal: true
 
+require "attrable"
 require "mutex_m"
 require "ignorable"
 
@@ -74,7 +75,7 @@ module Debrew
   @debugged_exceptions = Set.new
 
   class << self
-    extend Predicable
+    extend Attrable
     attr_predicate :active?
     attr_reader :debugged_exceptions
   end
